YAAAAAAAAAAAY BETTER JUMPING
Maybe. Hopefully. Probably. There definitelly won't have to be a change later and there definitelly won't be thousands of errors. Like a 90% chance. That that won't happen. Right? Idk. Anyway to make it so that you can jump smaller which kinda would have creted a lot of glitches I made it so that the initial liftoff is interpolated(I dont really understand how it works i just looked it up) to make it more smooth with still allowing you to do smaller jumps and not screwing up the jump when you want to hold. Yep. Surprised that this actually worked.