Fixed Wall Jumping
It is no longer weird. Basically it was pulling whether you could walljump from the wallcheck.position which gets the center of mass. Since the way they were positioned the center of mass was too high up causing you to not be able to jump sometimes. I also made wall jumping lerp because why not. It feels super smooth.